backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.35
python-multipart>=0.0.9
pillow>=11.0
pydantic>=2.8
sqlalchemy>=2.0
psycopg[binary]>=3.2
asyncpg>=0.30
pgvector>=0.3.6
aiofiles>=24.1
pyjwt>=2.10
httpx>=0.28
boto3>=1.36
twilio>=9.6
torch>=2.6
torchvision>=0.21
opencv-python-headless>=4.10
numpy>=1.26
transformers>=4.49
